instead of artus we now use excalibur (or if wanted also analysis wrapper could be used, but not recomended)
