--- BuildFile.xml
@@ -1,30 +1,0 @@
-<Flags LDFLAGS="-rdynamic" />
-
-
-<Flags LDFLAGS="-rdynamic" />
-<Flags LDFLAGS="-Lsrc/Kappa/lib -lKappa" />
-
-<use name="root" />
-<use name="tauspinner" />
-<use name="rootmath" />
-<use name="roofit"/>
-<use name="boost" />
-<use name="boost_regex" />
-<use name="CondFormats/JetMETObjects" />
-<use name="Artus/Filter" />
-<use name="Artus/KappaAnalysis" />
-<use name="Artus/Utility" />
-<use name="KappaTools/Toolbox" />
-<use name="KappaTools/RootTools" />
-<use name="Artus/Configuration"/>
-<use name="Artus/Consumer"/>
-<use name="Artus/Utility"/>
-
-
-<flags ADD_SUBDIR="1"/>
-
-<export>
-   <lib   name="1"/>
-</export>
-
-
