+++ Compile/bin/BuildFile.xml
@@ -1,0 +1,21 @@
+<bin name="excalibur" file="excalibur.cc">
+  <Flags LDFLAGS="-rdynamic" />
+  <use name="Excalibur/Compile" />
+  <use name="Kappa/DataFormats" />
+  <use name="root" />
+  <use name="rootmath" />
+  <use name="roofit"/>
+  <use name="boost" />
+  <use name="boost_regex" />
+  <use name="CondFormats/JetMETObjects" />
+  <use name="Artus/Filter" />
+  <use name="Artus/KappaAnalysis" />
+  <use name="KappaTools/Toolbox" />
+  <use name="KappaTools/RootTools" />
+  <use name="Artus/Configuration"/>
+  <use name="Artus/Consumer"/>
+  <use name="Artus/Utility"/>
+
+</bin>
+
+
