+++ Compile/interface/Consumers/ZJetTreeConsumer.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include "Artus/KappaAnalysis/interface/Consumers/KappaLambdaNtupleConsumer.h"
+
+#include "Excalibur/Compile/interface/ZJetTypes.h"
+
+class ZJetTreeConsumer : public KappaLambdaNtupleConsumer<ZJetTypes>
+{
+  public:
+    std::string GetConsumerId() const override;
+
+    void Init(ZJetSettings const& settings) override;
+};
