--- interface/Producers/ValidZllJetsProducer.h
@@ -1,20 +1,0 @@
-#pragma once
-
-#include "ZJetTypes.h"
-
-class ValidZllJetsProducer : public ZJetProducerBase
-{
-  public:
-    ValidZllJetsProducer() : ZJetProducerBase(), minZllJetDeltaRVeto(0) {}
-
-    void Init(ZJetSettings const& settings) override;
-
-    std::string GetProducerId() const override;
-
-    void Produce(ZJetEvent const& event,
-                 ZJetProduct& product,
-                 ZJetSettings const& settings) const override;
-
-  private:
-    float minZllJetDeltaRVeto;
-};
